LayoutTests/canvas/philip/tests/2d.state.saverestore.fillStyle-expected.txt

< [index] >
2d.state.saverestore.fillStyle
save()/restore() works for fillStyle
References: 2d.state.fillStyle
Actual output:
Passed
